Better controls in dumb sheet. Unified dumb common components
